renames Channel, Row; updates imports
