+++ src/components/editor/controls/SequencerChannel/SequencerChannel.css
