+++ src/components/editor/controls/SequencerRow/SequencerRow.css
