+++ src/components/editor/controls/index.js
@@ -2,7 +2,6 @@
 
 import AddTrack from './AddTrack/AddTrack.jsx';
 import BPMCounter from './BPMCounter/BPMCounter.jsx';
-import Channel from './Channel/Channel.jsx';
 import DeleteTrack from './DeleteTrack/DeleteTrack.jsx';
 import Dial from './Dial/Dial.jsx';
 import Dropdown from './Dropdown/Dropdown.jsx';
@@ -14,15 +13,15 @@ import KeyController from './KeyController/KeyController.jsx';
 import Pad from './Pad/Pad.jsx';
 import PianoKey from './PianoKey/PianoKey.jsx';
 import PlayButton from './PlayButton/PlayButton.jsx';
+import SequencerChannel from './SequencerChannel/SequencerChannel.jsx';
+import SequencerRow from './SequencerRow/SequencerRow.jsx';
 import Step from './Step/Step.jsx';
-import Row from './Row/Row.jsx';
 import VolumeFader from './VolumeFader/VolumeFader.jsx';
 import GlobalControls from './GlobalControls/GlobalControls.jsx';
 
 export {
   AddTrack,
   BPMCounter,
-  Channel,
   DeleteTrack,
   Dial,
   Dropdown,
@@ -34,8 +33,9 @@ export {
   Pad,
   PianoKey,
   PlayButton,
+  SequencerChannel,
+  SequencerRow,
   Step,
-  Row,
   VolumeFader,
   GlobalControls,
 };
